セレクトを使えるよう修正
