+++ client/pages/index.vue
@@ -18,7 +18,7 @@
 <script lang="ts">
 import { Component, Vue } from 'nuxt-property-decorator'
 import gql from 'graphql-tag'
-import { CONTACT_CATEGORIES } from '../utils'
+import { CONTACT_CATEGORIES } from '~/utils/contact'
 const MainTemplate = () => import('../components/layout/MainTemplate.vue')
 const ContactList = () => import('../components/contact/List.vue')
 const Pagination = () => import('../components/layout/Pagination.vue')
